feat(bare): migration metro to bare
