+++ docs/docs/guide/getting-started/quick-start-with-supabase.mdx
@@ -44,7 +44,7 @@ Run the initialization script to start the interactive setup process. Use the fo
 
 ::: details {title="Interactive Setup Steps"}
 
-1. **Select a Build Plugin**: Choose a build plugin for your project (e.g., Metro for React Native).
+1. **Select a Build Plugin**: Choose a build plugin for your project (e.g., Bare (React Native CLI) for React Native).
 2. **Select a Provider**: Select Supabase as the provider for handling updates.
 
 During the setup, you will be prompted to:
@@ -79,13 +79,13 @@ During the initialization process, the following file is automatically generated
 - `hot-updater.config.ts`: This file contains the configuration settings for integrating Supabase with your project.
 
 ```ts title=hot-updater.config.ts
-import { metro } from "@hot-updater/metro";
+import { bare } from "@hot-updater/bare";
 import { supabaseDatabase, supabaseStorage } from "@hot-updater/supabase";
 import { defineConfig } from "hot-updater";
 import "dotenv/config";
 
 export default defineConfig({
-  build: metro({ enableHermes: true }),
+  build: bare({ enableHermes: true }),
   storage: supabaseStorage({
     supabaseUrl: process.env.HOT_UPDATER_SUPABASE_URL!,
     supabaseAnonKey: process.env.HOT_UPDATER_SUPABASE_ANON_KEY!,
