+++ docs/docs/guide/plugin-system.mdx
@@ -11,8 +11,7 @@ The **Build Plugin** is used during the execution of the `hot-updater deploy` co
 
 ### Supported Plugins
 
-- **`@hot-updater/metro`**: Supports Metro bundler for React Native projects.
-- **`@hot-updater/repack`**: Coming soon.
+- **`@hot-updater/bare`**: Executes the bundle command through React Native CLI. Bundles through Metro or Re.Pack.
 
 ## Storage Plugin
