+++ docs/docs/guide/providers/2_cloudflare.mdx
@@ -41,7 +41,7 @@ Run the initialization script to start the interactive setup process. Use the fo
 
 ::: details {title="Interactive Setup Steps"}
 
-1. **Select a Build Plugin**: Choose a build plugin for your project (e.g., Metro for React Native).
+1. **Select a Build Plugin**: Choose a build plugin for your project (e.g., Bare (React Native CLI) for React Native).
 2. **Select a Provider**: Select Cloudflare as the provider for handling updates.
 
 During the setup, you will be prompted to:
@@ -80,13 +80,13 @@ During the initialization process, the following file is automatically generated
 - `hot-updater.config.ts`: This file contains the configuration settings for integrating Cloudflare with your project.
 
 ```ts title=hot-updater.config.ts
-import { metro } from "@hot-updater/metro";
+import { bare } from "@hot-updater/bare";
 import { d1Database, r2Storage } from "@hot-updater/cloudflare";
 import { defineConfig } from "hot-updater";
 import "dotenv/config";
 
 export default defineConfig({
-  build: metro({ enableHermes: true }),
+  build: bare({ enableHermes: true }),
   storage: r2Storage({
     bucketName: process.env.HOT_UPDATER_CLOUDFLARE_R2_BUCKET_NAME!,
     accountId: process.env.HOT_UPDATER_CLOUDFLARE_ACCOUNT_ID!,
