+++ examples/v0.71.19/hot-updater.config.ts
@@ -1,4 +1,4 @@
-import { metro } from "@hot-updater/metro";
+import { bare } from "@hot-updater/bare";
 import { withSentry } from "@hot-updater/sentry-plugin";
 import { supabaseDatabase, supabaseStorage } from "@hot-updater/supabase";
 import { defineConfig } from "hot-updater";
@@ -6,7 +6,7 @@ import "dotenv/config";
 
 export default defineConfig({
   build: withSentry(
-    metro({
+    bare({
       enableHermes: false,
       sourcemap: true, // sentry will upload sourcemaps to sentry
     }),
