+++ examples/v0.74.1/package.json
@@ -25,7 +25,7 @@
     "@babel/runtime": "^7.24.5",
     "@callstack/repack": "^4.0.0",
     "@hot-updater/aws": "0.16.3",
-    "@hot-updater/metro": "0.16.3",
+    "@hot-updater/bare": "0.16.3",
     "@hot-updater/supabase": "0.16.3",
     "@react-native-community/cli-platform-android": "13.6.6",
     "@react-native/babel-preset": "0.74.83",
