+++ examples/v0.76.1-new-arch/hot-updater.config.ts
@@ -1,10 +1,10 @@
-import { metro } from "@hot-updater/metro";
+import { bare } from "@hot-updater/bare";
 import { supabaseDatabase, supabaseStorage } from "@hot-updater/supabase";
 import { defineConfig } from "hot-updater";
 import "dotenv/config";
 
 export default defineConfig({
-  build: metro({
+  build: bare({
     enableHermes: true,
   }),
   storage: supabaseStorage({
