+++ examples/v0.77.0/hot-updater.config.ts
@@ -1,10 +1,10 @@
-import {metro} from '@hot-updater/metro';
-import {supabaseDatabase, supabaseStorage} from '@hot-updater/supabase';
-import {defineConfig} from 'hot-updater';
-import 'dotenv/config';
+import { bare } from "@hot-updater/bare";
+import { supabaseDatabase, supabaseStorage } from "@hot-updater/supabase";
+import { defineConfig } from "hot-updater";
+import "dotenv/config";
 
 export default defineConfig({
-  build: metro({enableHermes: true}),
+  build: bare({ enableHermes: true }),
   storage: supabaseStorage({
     supabaseUrl: process.env.HOT_UPDATER_SUPABASE_URL!,
     supabaseAnonKey: process.env.HOT_UPDATER_SUPABASE_ANON_KEY!,
